Use correct type for spi mode

+++ src/device/device_spi.cpp
@@ -19,7 +19,7 @@ control_ret_t Device::device_init()
     control_ret_t ret = CONTROL_SUCCESS;
     if(!device_initialised)
     {
-        ret = control_init_spidev(static_cast<int>(device_info[0]),       // SPI mode
+        ret = control_init_spidev(static_cast<uint8_t>(device_info[0]),   // SPI mode
                                   static_cast<uint32_t>(device_info[1]),  // SPI frequency
                                   static_cast<int>(device_info[2]),       // spidev bus
                                   static_cast<int>(device_info[3]),       // spidev cs
